Add greater configuration in target dsl
This allows the project owner to specify the strictness of type
checks in their project. The naming of the new options draws upon
the structure of similar options in mypy[0] ang the naming of the
typing errors themselves.

These two errors are very common ones that arise when trying to add
steep to an existing untyped project.

[0] https://mypy.readthedocs.io/en/stable/config_file.html