completion: Support completion for optional chaining (&.)
The optional chaining allows to developers to call method only if the
receiver is not nil.  Therefore, CompletionProvider should consider the
receiver is not nil if the optional chaining operator is given.

This fix unwraps the Optional[T] from receiver_type on completion for
the optional chaining.  It will provide better completion for
developers.