Version 1.5.0.pre.2
The last release `1.5.0.pre.1` didn’t contain #823 because it was not synchronized with the `HEAD` while release work…

The `1.5.0.pre.1` was released from my laptop and the commit history of `master` is incorrect.

`1.5.0.pre.2` fixes the problem. The release is synchronized with the commit on `master` and CHANGELOG was fixed.