langserver: Add --severity-level option
Currently, langserver does not support `--severity-level` option.
Therefore the diagnostics are different with `steep check` command
by default because the result of `steep check` is filtered by
severity-level warning by default.  On the other hand, the result of
`steep langserver` is not filtered at all.

This adds `--severity-level` option to the `steep langserver` command to
control the severity level through the option.  And it defaults to
"warning" level as same as other subcommand does.