Add a new diagnostic type: NoMethodByNil
Add a new diagnostic type; NoMethodByNil, that is emitted to no method
calls for the objects having optional types.  It's separated from
NoMethod to control the visiblity of these diagnostics individually.

It's useful to find the methods not having appropriate types on checking
the Ruby application by filtering NoMethod diagnostics.