Adjust LSP test
The test hinges on a plain `attr_reader` going from being annotated to
unannotated.

Before the change it went from a non-error state to an error state,
which is used as a criteria for the LSP to

After the change the annotation is not required anymore, therefore the
LSP test endlessly waits for an error that never occurs.

By moving to another type of error the LSP test now passes again.

In addition the annotation was removed as it is unneeded now.