Type narrowing union types via return type of method call
This tries to narrow types unions by the return type of the method call.
If the return type of the method is always truthy or falsy, the receiver
type will be narrowed on the truthy branch and the falsy branch.

This helps to narrow types via method definitions (ex. `#nil?`,
`#present?`, `#blank?` and so on).